requirements.txt
# Web Framework Dependencies
fastapi==0.115.14
uvicorn==0.35.0
gunicorn==23.0.0
starlette==0.47.1
pydantic==2.11.7
pydantic-core==2.33.2

# Flask Dependencies
Flask==3.1.1
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2
itsdangerous==2.2.0

# AWS Lambda Integration
mangum==0.19.0

# HTTP Client Libraries
requests==2.32.4
httpx==0.28.1
httpcore==1.0.9
h11==0.16.0
urllib3==2.5.0

# Google AI/Gemini Dependencies
google-generativeai==0.8.5
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-api-core==2.25.1
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.71.2
proto-plus==1.26.1
protobuf==5.29.5

# Database Connectivity
mysql-connector-python==9.3.0

# Machine Learning/Embeddings
sentence-transformers
numpy
torch
transformers

# Environment and Configuration
python-dotenv==1.1.1

# Utility Libraries
click==8.2.1
tqdm==4.67.1
colorama==0.4.6
blinker==1.9.0
anyio==4.9.0
sniffio==1.3.1
idna==3.10
certifi==2025.6.15
charset-normalizer==3.4.2

# Cryptography and Security
rsa==4.9.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
cachetools==5.5.2

# Type Checking and Validation
typing-extensions==4.14.0
annotated-types==0.7.0

# Build Tools
setuptools==80.9.0
wheel==0.45.1

# Additional Utilities
uritemplate==4.2.0
httplib2==0.22.0
pyparsing==3.2.3